Mutation work
Added babel polyfill with no conflict for heroku reasons. Added the request property to the ApolloClient which takes the users token from the sessionStorage if there is one available. Made a little Mutation - login mutation test and observed, that the token is getting passed to the server via headers (which is soooo helpful!).